magic code

+++ Sources/Microcontroller Codes/feedback control/controller_lib_playground/infraredDirection.h
@@ -16,7 +16,7 @@ class infraredDirection
     bool init();
     bool update();
     bool isVisible();
-    int getDistance();
+    float getDistance();
     float getDirection();
     float getRightVoltage();
     float getLeftVoltage();
